Update Iris benchmarking to align with templating.
