+++ benchmarks/asv.conf.json
@@ -3,7 +3,7 @@
     "project": "scitools-iris",
     "project_url": "https://github.com/SciTools/iris",
     "repo": "..",
-    "environment_type": "delegated-iris",
+    "environment_type": "delegated",
     "show_commit_url": "https://github.com/scitools/iris/commit/",
     "branches": ["upstream/main"],
 
@@ -11,7 +11,7 @@
     "env_dir": ".asv/env",
     "results_dir": ".asv/results",
     "html_dir": ".asv/html",
-    "plugins": [".asv_delegated_iris"],
+    "plugins": [".asv_delegated"],
 
     "command_comment": [
         "The inherited setup of the Iris test environment takes care of ",
